fix(server): PC2 line-check — extract ci-event handler to server_ci_route.c
server_http_routes.c exceeded the 2500-line limit after the CI webhook. Move
ci_hmac_sha256_hex + rh_dev_ci_event to a new server/server_ci_route.c; share
route_req_t/route_handler_fn + declare rh_dev_ci_event via server_http_internal.h.
routes.c now 2493 lines. Test link + SERVER_SRCS updated.
